+++ src/main/java/cz/craftmania/crafteconomy/sql/SQLManager.java
@@ -606,7 +606,7 @@ public class SQLManager {
         }.runTaskAsynchronously(Main.getInstance());
     }
 
-    public final long getVaultEcoBalance(final UUID uuid) {
+    public final double getVaultEcoBalance(final UUID uuid) {
         final String server = Main.getServerType().name().toLowerCase();
         Connection conn = null;
         PreparedStatement ps = null;
@@ -615,7 +615,7 @@ public class SQLManager {
             ps = conn.prepareStatement("SELECT balance FROM player_economy_" + server + " WHERE uuid = '" + uuid + "';");
             ps.executeQuery();
             if (ps.getResultSet().next()) {
-                return ps.getResultSet().getLong("balance");
+                return ps.getResultSet().getDouble("balance");
             }
         } catch (Exception e) {
             Main.getInstance().sendSentryException(e);
@@ -626,7 +626,7 @@ public class SQLManager {
         return 0;
     }
 
-    public final long getVaultEcoBalance(final String nick) {
+    public final double getVaultEcoBalance(final String nick) {
         final String server = Main.getServerType().name().toLowerCase();
         Connection conn = null;
         PreparedStatement ps = null;
@@ -635,7 +635,7 @@ public class SQLManager {
             ps = conn.prepareStatement("SELECT balance FROM player_economy_" + server + " WHERE nick = '" + nick + "';");
             ps.executeQuery();
             if (ps.getResultSet().next()) {
-                return ps.getResultSet().getLong("balance");
+                return ps.getResultSet().getDouble("balance");
             }
         } catch (Exception e) {
             Main.getInstance().sendSentryException(e);
@@ -646,8 +646,8 @@ public class SQLManager {
         return 0;
     }
 
-    public Map<String, Long> getVaultAllEcosWithNicks() {
-        Map<String, Long> balanceMap = new HashMap<String, Long>();
+    public Map<String, Double> getVaultAllEcosWithNicks() {
+        Map<String, Double> balanceMap = new HashMap<String, Double>();
 
         final String server = Main.getServerType().name().toLowerCase();
         Connection conn = null;
@@ -658,7 +658,7 @@ public class SQLManager {
             ps = conn.prepareStatement("SELECT `nick`, `balance` FROM `player_economy_" + server + "` WHERE `balance` > 0 AND `hide_in_baltop` = 0 ORDER BY `balance` DESC");
             ps.executeQuery();
             while (ps.getResultSet().next()) {
-                balanceMap.put(ps.getResultSet().getString("nick"), ps.getResultSet().getLong("balance"));
+                balanceMap.put(ps.getResultSet().getString("nick"), ps.getResultSet().getDouble("balance"));
             }
         } catch (Exception e) {
             Main.getInstance().sendSentryException(e);
@@ -733,7 +733,7 @@ public class SQLManager {
                     if (resultSet.getString("s_uuid") == null) senderUUID = null;
                     else senderUUID = UUID.fromString(resultSet.getString("s_uuid"));
                     final EconomyLog.EconomyAction action = EconomyLog.EconomyAction.valueOf(resultSet.getString("action"));
-                    final Long amount = resultSet.getLong("amount");
+                    final double amount = resultSet.getLong("amount");
                     final Long time = resultSet.getLong("time");
 
                     EconomyLog economyLog = new EconomyLog(reciever, recieverUUID, sender, senderUUID, action, amount, time);
@@ -778,7 +778,7 @@ public class SQLManager {
                     if (resultSet.getString("s_uuid") == null) senderUUID = null;
                     else senderUUID = UUID.fromString(resultSet.getString("s_uuid"));
                     final EconomyLog.EconomyAction action = EconomyLog.EconomyAction.valueOf(resultSet.getString("action"));
-                    final Long amount = resultSet.getLong("amount");
+                    final double amount = resultSet.getLong("amount");
                     final Long time = resultSet.getLong("time");
 
                     EconomyLog economyLog = new EconomyLog(reciever, recieverUUID, sender, senderUUID, action, amount, time);
@@ -819,7 +819,7 @@ public class SQLManager {
         throw new NullPointerException("UUID pro hrace (" + name + ") nebylo nalezeno!");
     }
 
-    public void setVaultEcoBalance(final String player, final long amount) {
+    public void setVaultEcoBalance(final String player, final double amount) {
         Main.getAsync().runAsync(() -> {
             long milis = System.currentTimeMillis();
             final String server = Main.getServerType().name().toLowerCase();
@@ -828,7 +828,7 @@ public class SQLManager {
             try {
                 conn = pool.getConnection();
                 ps = conn.prepareStatement("UPDATE player_economy_" + server + " SET balance = ?, last_update = ? WHERE nick = ?");
-                ps.setLong(1, amount);
+                ps.setDouble(1, amount);
                 ps.setLong(2, System.currentTimeMillis());
                 ps.setString(3, player);
                 ps.executeUpdate();
@@ -856,7 +856,7 @@ public class SQLManager {
                     ps = conn.prepareStatement("INSERT INTO player_economy_" + server + " (nick, uuid, balance, last_update) VALUES (?,?,?,?);");
                     ps.setString(1, player.getName());
                     ps.setString(2, player.getUniqueId().toString());
-                    ps.setLong(3, startValue);
+                    ps.setDouble(3, startValue);
                     ps.setLong(4, currentTime);
                     ps.executeUpdate();
                 } catch (Exception e) {
@@ -909,8 +909,8 @@ public class SQLManager {
         }
     }
 
-    public final List<Triple<String, Long, Long>> fetchAllEconomyRowsToRemove(long time) {
-        List<Triple<String, Long, Long>> balanceList = new ArrayList<>();
+    public final List<Triple<String, Double, Long>> fetchAllEconomyRowsToRemove(long time) {
+        List<Triple<String, Double, Long>> balanceList = new ArrayList<>();
         final String server = Main.getServerType().name().toLowerCase();
         Connection conn = null;
         PreparedStatement ps = null;
@@ -923,7 +923,7 @@ public class SQLManager {
                 balanceList.add(
                         new Triple<>(
                                 ps.getResultSet().getString("nick"),
-                                ps.getResultSet().getLong("balance"),
+                                ps.getResultSet().getDouble("balance"),
                                 ps.getResultSet().getLong("last_update"))
                         );
             }
